Relocate Snowflake Type
