+++ packages/interaction-kit/src/interactions/application-commands/application-command-interaction.ts
@@ -6,8 +6,8 @@ import {
 	InteractionReply,
 	RequestBody,
 	ResponseHandler,
-	Snowflake,
 } from "../../interfaces";
+import type { Snowflake } from "../../structures/snowflake";
 import {
 	APIApplicationCommandInteraction,
 	APIInteractionGuildMember,
